copyright removed (is in a separate file) remove appveyor also test material movements sql scripts to own files refactoring allow first migration to fail invoice test on VAT EUR using the new client 1000000 run against customized iDempiere 5.1 client setup (works) run against customized iDempiere 5.1 client setup invoicing test fixed DefaultDocumentFactory [WIP] tests work

+++ src/main/java/org/compiere/order/X_C_Order.java
@@ -585,7 +585,7 @@ public class X_C_Order extends PO implements I_Persistent
 		if (C_DocTypeTarget_ID < 1) 
 			set_Value (I_C_Order.COLUMNNAME_C_DocTypeTarget_ID, null);
 		else 
-			set_Value (I_C_Order.COLUMNNAME_C_DocTypeTarget_ID, Integer.valueOf(C_DocTypeTarget_ID));
+			set_Value (I_C_Order.COLUMNNAME_C_DocTypeTarget_ID, C_DocTypeTarget_ID);
 	}
 
 	/** Get Target Document Type.
@@ -596,7 +596,7 @@ public class X_C_Order extends PO implements I_Persistent
 		Integer ii = (Integer)get_Value(I_C_Order.COLUMNNAME_C_DocTypeTarget_ID);
 		if (ii == null)
 			 return 0;
-		return ii.intValue();
+		return ii;
 	}
 
 	/** Set Charge amount.
